Add -a awk option
New script version 1.0.2 2023-07-09 (even though the month and date are
backwards :-) )

This commit also make it so the script checks that awk exists, is a
regular file and is executable whereas before it just assumed it was
found. Script exits 4 if any of these conditions fail (that is it does
not exist, is not a regular file or is not an executable file).