Check that fmt and awk exist prior to use
New version 1.0.3 2023-07-11.

Add missing check for awk existing. Due to a problem I was having
initially (with my fork) the check was not put in place by accident.
It's good that this happened because there was a bug where if one did
not specify the option it would have ended up being an error since if
the option is not specified we have to find the path first before we
check for the existence of it.

In particular if -a awk or -f fmt is/are not used then we use type -P on
the string or strings (awk / fmt) to get the correct path. After that we
can do a test on the files.

The script now also checks for fmt prior to use in the same way that it
does for awk.