Fix mix-up in exit code check for debug statement
The script reported exit code if it was not equal to 1 rather than if it
was not equal to 0 and I believe this is a mistake as it should report
if it's an error rather than success. This was not shown by default as
it only happens if the verbosity is >= 1.

Typo fixed statement too: ext -> exit.